refac: start using org.json for easier rest template data processing

+++ src/main/java/com/doddysujatmiko/rumiapi/anime/AnimeEntity.java
@@ -24,7 +24,7 @@ public class AnimeEntity extends BaseEntity {
 
     private String japaneseTitle;
 
-    private double rating;
+    private float rating;
 
     @Column(length = 2048)
     private String description;
